Implement a pseudocode generator (#5)
This commit implements a pseudocode generator which can be used to convert an already serialized Section data structure back into bygg C++ code. Coupled with the parser, you can more or less convert documents back into bygg syntax.

It is, of course, pseudocode and may not necessarily compile nor be best practice, but it can be a time saver. Generally, simple structures should be a drop in replacement.